app/requirements.txt
Django==5.0.6
celery==5.4.0
django-celery-beat==2.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-django==4.8.0
PyJWT==2.3.0
